친구 관련 컴포넌트 재사용을 위한 컴포넌트로 분리
